docs: 파일 정리
